Debugging and MWEs

There's a reasonable chance that you'll run into an issue with Tapir.jl at some point. In order to debug what is going on when this happens, or to produce an MWE, it is helpful to have a convenient way to run Tapir.jl on whatever function and arguments you have which are causing problems.

We recommend making use of Tapir.jl's testing functionality to generate your test cases:

Tapir.TestUtils.test_rule — Function

test_rule(
    rng, x...;
    interface_only=false,
    has_rrule!!=true,
    perf_flag::Symbol,
    ctx=DefaultCtx(),
    safety_on=false,
)

Run standardised tests on the rule for x. The first element of x should be the primal function to test, and each other element a positional argument. In most cases, elements of x can just be the primal values, and randn_tangent can be relied upon to generate an appropriate tangent to test. Some notable exceptions exist though, in partcular Ptrs. In this case, the argument for which randn_tangent cannot be readily defined should be a CoDual containing the primal, and a manually constructed tangent field.

This function uses Tapir.build_rrule to construct a rule. This will use an rrule!! if one exists, and derive a rule otherwise.

This approach is convenient because it can

check whether AD runs at all,

check whether AD produces the correct answers,

check whether AD is performant, and

can be used without having to manually generate tangents.

Example

For example

f(x) = Core.bitcast(Float64, x)
Tapir.TestUtils.test_rule(
    Random.Xoshiro(123), f, 3;
    is_primitive=false, perf_flag=:none, interp=Tapir.TapirInterpreter(),
)

will error. (In this particular case, it is caused by Tapir.jl preventing you from doing (potentially) unsafe casting. In this particular instance, Tapir.jl just fails to compile, but in other instances other things can happen.)

In any case, the point here is that Tapir.TestUtils.test_rule provides a convenient way to produce and report an error.

Segfaults

These are everyone's least favourite kind of problem, and they should be extremely rare in Tapir.jl. However, if you are unfortunate enough to encounter one, please re-run your problem with the safety_on kwarg set to true. See Safe Mode for more info. In general, this will catch problems before they become segfaults, at which point the above strategy for debugging and error reporting should work well.