Tapir.jl's Mathematical Intepretation of Julia Functions

A Mathematical Model for a Computer Programme

In order to make sense of what it might mean to differentiate a computer programme, we need some kind of differentiable mathematical model for said programme. Pearlmutter [3] introduced such a model, which we expand on here.

Think of a computer as a device which has a state, which we model as a vector in $\RR^D$. When a programme p is run, it changes this state. We model the programme p with a "transition function" $t : \RR^D \to \RR^D$, whose input is whatever the state is before running p, and whose output is whatever the state is after running p. If $t$ is differentiable, then we can meaningfully define "differentiating p" to mean "differentiating $t$". In particular, when we talk about applying reverse-mode AD to a particular Julia function p, what we mean is computing the adjoint $D t [x]$ of the transition function $t$ which describes p when the computer is in state $x$ prior to running p.

Consider how we might model the Julia function

function f(x)
    x1 = x
    x2 = f1(x1)
    x3 = f2(x1, x2)
    ...
    return fN(x1, x2, ..., xN)
end

It can be thought of as a sequence of $N$ programmes, which are run one after the other. If each fn is associated to transition functin $t_n : \RR^D \to \RR^D$ then the transtion function associated to f is just $t := t_N \circ \dots \circ t_1$. We can then compute the adjoint of $t$ in the way described in the last section:

\[D t [x]^\ast (\bar{y}) = [ D t_1 [x_1] \circ \dots \circ D t_N [x_N] ]^\ast (\bar{y})\]

A Simple Worked Example

Consider applying this framework to the Julia function

f(x::Float64) = sin(cos(x))

We first find the transition operator associated to the primitives sin(::Float64) and cos(::Float64), then look at

Advantages and Limitations

In principle, we have now associated a precise meaning to applying reverse-mode AD to the Julia function f and shown how reverse-mode AD might be applied. This view has several advantages. It places sufficiently few restrictions on what a computer programme is allowed to do that we can imagine it being a good model for most computer programmes we shall encounter. Moreover, it gives us a precise meaning to "differentiating a computer programme".

However, it has returned us to working with "flat" vector structures, rather than directly with the

Tangents

We call the argument or output of a derivative $D f [x] : \mathcal{X} \to \mathcal{Y}$ a tangent, and will usually denote it with a dot over a symbol, e.g. $\dot{x}$. Conversely, we call an argument or output of the adjoint of this derivative $D f [x]^\ast : \mathcal{Y} \to \mathcal{X}$ a gradient, and will usually denote it with a bar over a symbol, e.g. $\bar{y}$.

Note, however, that the sets involved are the same whether dealing with a derivative or its adjoint. Consequently, we use the same type to represent both.

A quick aside: Non-Differentiable Data

In the introduction to algorithmic differentiation, we assumed that the domain / range of function are the same as that of its derivative. Unfortunately, this story is only partly true. Matters are complicated by the fact that not all data types in Julia can reasonably be thought of as forming a Hilbert space. e.g. the String type.

Consequently we introduce the special type NoTangent, instances of which can be thought of as representing the set containing only a $0$ tangent. Morally speaking, for any non-differentiable data x, x + NoTangent() == x.

Other than non-differentiable data, the model of data in Julia as living in a real-valued finite dimensional Hilbert space is quite reasonable. Therefore, we hope readers will forgive us for largely ignoring the distinction between the domain and range of a function and that of its derivative in mathematical discussions, while simultaneously drawing a distinction when discussing code.

Representing Tangents

The extended docstring for tangent_type provides the best introduction to the types which are used to represent tangents.

Tapir.tangent_type — Method

tangent_type(P)

There must be a single type used to represents tangents of primals of type P, and it must be given by tangent_type(P).

Extended help

The tangent types which Tapir.jl uses are quite similar in spirit to ChainRules.jl. For example, tangent "vectors" for

Float64s are Float64s,

Vector{Float64}s are Vector{Float64}s, and

structs are other another (special) struct with field types specified recursively.

There are, however, some major differences. Firstly, while it is certainly true that the above tangent types are permissible in ChainRules.jl, they are not the uniquely permissible types. For example, ZeroTangent is also a permissible type of tangent for any of them, and Float32 is permissible for Float64. This is a general theme in ChainRules.jl – it intentionally declines to place restrictions on what type can be used to represent the tangent of a given type.

Tapir.jl differs from this. It insists that each primal type is associated to a single tangent type. Furthermore, this type is always given by the function Tapir.tangent_type(primal_type).

Consider some more worked examples.

Int

Int is not a differentiable type, so its tangent type is NoTangent:

julia> tangent_type(Int)
NoTangent

Tuples

The tangent type of a Tuple is defined recursively based on its field types. For example

julia> tangent_type(Tuple{Float64, Vector{Float64}, Int})
Tuple{Float64, Vector{Float64}, NoTangent}

There is one edge case to be aware of: if all of the field of a Tuple are non-differentiable, then the tangent type is NoTangent. For example,

julia> tangent_type(Tuple{Int, Int})
NoTangent

Structs

As with Tuples, the tangent type of a struct is, by default, given recursively. In particular, the tangent type of a struct type is Tangent. This type contains a NamedTuple containing the tangent to each field in the primal struct.

As with Tuples, if all field types are non-differentiable, the tangent type of the entire struct is NoTangent.

There are a couple of additional subtleties to consider over Tuples though. Firstly, not all fields of a struct have to be defined. Fortunately, Julia makes it easy to determine how many of the fields might possibly not be defined. The tangent associated to any field which might possibly not be defined is wrapped in a PossiblyUninitTangent.

Furthermore, structs can have fields whose static type is abstract. For example

julia> struct Foo
           x
       end

If you ask for the tangent type of Foo, you will see that it is

julia> tangent_type(Foo)
Tangent{@NamedTuple{x}}

Observe that the field type associated to x is Any. The way to understand this result is to observe that

x could have literally any type at runtime, so we know nothing about what its tangent type must be until runtime, and

we require that the tangent type of Foo be unique.

The consequence of these two considerations is that the tangent type of Foo must be able to contain any type of tangent in its x field. It follows that the fieldtype of the x field of Foos tangent must be Any.

Mutable Structs

The tangent type for mutable structs have the same set of considerations as structs. The only difference is that they must themselves be mutable. Consequently, we use a type called MutableTangent to represent their tangents. It is a mutable struct with the same structure as Tangent.

For example, if you ask for the tangent_type of

julia> mutable struct Bar
           x::Float64
       end

you will find that it is

julia> tangent_type(Bar)
MutableTangent{@NamedTuple{x::Float64}}

Primitive Types

We've already seen a couple of primitive types (Float64 and Int). The basic story here is that all primitive types require an explicit specification of what their tangent type must be.

One interesting case are Ptr types. The tangent type of a Ptr{P} is Ptr{T}, where T = tangent_type(P). For example

julia> tangent_type(Ptr{Float64})
Ptr{Float64}

FData and RData

While tangents are the things used to represent gradients, they are not strictly what gets propagated forwards and backwards by rules during AD. Rather, they are split into fdata and rdata, and these are passed around.

Tapir.fdata_type — Method

fdata_type(T)

Returns the type of the forwards data associated to a tangent of type T.

Extended help

Rules in Tapir.jl do not operate on tangents directly. Rather, functionality is defined to split each tangent into two components, that we call fdata (forwards-pass data) and rdata (reverse-pass data). In short, any component of a tangent which is identified by its address (e.g. a mutable structs or an Array) gets passed around on the forwards-pass of AD and is incremented in-place on the reverse-pass, while components of tangents identified by their value get propagated and accumulated only on the reverse-pass.

Given a tangent type T, you can find out what type its fdata and rdata must be with fdata_type(T) and rdata_type(T) respectively. A consequence of this is that there is exactly one valid fdata type and rdata type for each primal type.

Given a tangent t, you can get its fdata and rdata using f = fdata(t) and r = rdata(t) respectively. f can be re-combined to recover the original tangent using the binary version of tangent: tangent(f, r). It must always hold that

tangent(fdata(t), rdata(t)) === t

The need for all of this is explained in the docs, but for now it suffices to consider our running examples again, and to see what their fdata and rdata look like.

Int

Ints are non-differentiable types, so there is nothing to pass around on the forwards- or reverse-pass. Therefore

julia> fdata_type(tangent_type(Int)), rdata_type(tangent_type(Int))
(NoFData, NoRData)

Float64

The tangent type of Float64 is Float64. Float64s are identified by their value / have no fixed address, so

julia> (fdata_type(Float64), rdata_type(Float64))
(NoFData, Float64)

Vector{Float64}

The tangent type of Vector{Float64} is Vector{Float64}. A Vector{Float64} is identified by its address, so

julia> (fdata_type(Vector{Float64}), rdata_type(Vector{Float64}))
(Vector{Float64}, NoRData)

Tuple{Float64, Vector{Float64}, Int}

This is an example of a type which has both fdata and rdata. The tangent type for Tuple{Float64, Vector{Float64}, Int} is Tuple{Float64, Vector{Float64}, NoTangent}. Tuples have no fixed memory address, so we interogate each field on its own. We have already established the fdata and rdata types for each element, so we recurse to obtain:

julia> T = tangent_type(Tuple{Float64, Vector{Float64}, Int})
Tuple{Float64, Vector{Float64}, NoTangent}

julia> (fdata_type(T), rdata_type(T))
(Tuple{NoFData, Vector{Float64}, NoFData}, Tuple{Float64, NoRData, NoRData})

The zero tangent for (5.0, [5.0]) is t = (0.0, [0.0]). fdata(t) returns (NoFData(), [0.0]), where the second element is === to the second element of t. rdata(t) returns (0.0, NoRData()). In this example, t contains a mixture of data, some of which is identified by its value, and some of which is identified by its address, so there is some fdata and some rdata.

Structs

Structs are handled in more-or-less the same way as Tuples, albeit with the possibility of undefined fields needing to be explicitly handled. For example, a struct such as

julia> struct Foo
           x::Float64
           y
           z::Int
       end

has tangent type

julia> tangent_type(Foo)
Tangent{@NamedTuple{x::Float64, y, z::NoTangent}}

Its fdata and rdata are given by special FData and RData types:

julia> (fdata_type(tangent_type(Foo)), rdata_type(tangent_type(Foo)))
(Tapir.FData{@NamedTuple{x::NoFData, y, z::NoFData}}, Tapir.RData{@NamedTuple{x::Float64, y, z::NoRData}})

Practically speaking, FData and RData both have the same structure as Tangents and are just used in different contexts.

Mutable Structs

The fdata for a mutable structs is its tangent, and it has no rdata. This is because mutable structs have fixed memory addresses, and can therefore be incremented in-place. For example,

julia> mutable struct Bar
           x::Float64
           y
           z::Int
       end

has tangent type

julia> tangent_type(Bar)
MutableTangent{@NamedTuple{x::Float64, y, z::NoTangent}}

and fdata / rdata types

julia> (fdata_type(tangent_type(Bar)), rdata_type(tangent_type(Bar)))
(MutableTangent{@NamedTuple{x::Float64, y, z::NoTangent}}, NoRData)

Primitive Types

As with tangents, each primitive type must specify what its fdata and rdata is. See specific examples for details.

The Rule Abstraction

A rule must return a CoDual and a function to run the reverse-pass, known as the pullback. Upon exit from the rule, it must be true that

the state of the arguments / output are the same as they would be had the primal been run, and

the uniqueness of the mapping between address-identified primals and their fdata is maintained.

Upon exit from the pullback, it must be true that

the primal state is as it was before running the rule,

the fdata for the arguments has been incremented by the fdata in $D f[x]^\ast (\bar{y})$, and

the rdata for the arguments is equal to the rdata in $D f[x]^\ast (\bar{y})$.

Tapir.jl makes use of a rule system which is at first glance similar to the rrule function offered by ChainRules.jl. However, owing to Tapir.jl's support for mutation (e.g. differentiating through functions which write to arrays) and high degree of precision around the types used to represent (co)-tangent-like data, the number of situations in which the two are identical are actually rather small.

Nevertheless, we begin this explanation with an example which should be familiar to anyone who has used ChainRules.jl and seen its rrule. Once this example has been explained, we move into new territory.

Functions of Scalars: from ChainRules.rrule to Tapir.rrule!!

Consider the simple Julia function

mul(a::Float64, b::Float64) = a * b

A ChainRules.rrule for this might look something like

function ChainRules.rrule(::typeof(mul), a::Float64, b::Float64)
    mul_pullback(dc::Float64) = NoTangent(), dc * b, dc * a
    return a * b, mul_pullback
end

The corresponding Tapir.rrule!! would be something like

function Tapir.rrule!!(::CoDual{typeof(mul)}, a::CoDual{Float64}, b::CoDual{Float64})
    _a = primal(a)
    _b = primal(b)
    mul_pullback!!(dc::Float64) = NoRData(), dc * _b, dc * _a
    return CoDual(_a * _b, NoFData()), mul_pullback!
end

The core differences between the rrule and rrule!! are:

each argument is a CoDual, which contains the primal and one other piece of data (more on this later),

we must extract the primal values from a and b using the primal function in order to access them,

NoTangent() is replaced by NoRData(), and

we must return another CoDual, rather than just the primal value (more on this later).

The point of this example is to highlight that Tapir.rrule!!s look a lot like ChainRules.rrules in some situations, so some of your existing knowledge should transfer over.

Functions of Vectors

We now turn to the obvious question: why do Tapir.rrule!!s differ from ChainRules.rrules? The short answer is that Tapir.jl requires that each unique primal memory address associated to differentiable data be associated to a unique tangent (a.k.a. shadow) memory address. (See Why Unique Memory Address to understand why this is necessary.)

To see how this is achieved, consider the function

function set_1!(x::Vector{Float64}, y::Float64)
    x[1] = y
    return x
end

A valid Tapir.rrule!! for this function given below. There are a lot of concepts introduced here, so you'll need to hop back and forth between this and the text below which explains everything.

function Tapir.rrule!!(
    ::CoDual{typeof(set_1!)}, x::CoDual{Vector{Float64}}, y::CoDual{Float64}
)
    # Extract the primal and "fdata" from x.
    px = primal(x)
    dx = tangent(x)

    # Store the current values.
    px_1_old = px[1]
    dx_1_old = dx[1]

    # Set x_p[1] to `y` and zero-out x_f[1].
    px[1] = primal(y)
    dx[1] = 0.0

    function set_1_pullback!!(::NoRData)

        # The (co)tangent to `y` is just the value in the first position of x_f.
        dy = dx

        # We _must_ undo any changes which occur on the forwards-pass, both to the primal
        # and the fdata (the forwards-component of the tangent).
        px[1] = px_1_old
        dx[1] = dx_1_old

        # There's nothing to propagate backwards for `f` because it's non-differentiable.
        # It has "no reverse data", hence `NoRData`.
        df = NoRData()

        # There's nothing to propagate backwards for `x`, because its tangent is entirely
        # represented by `dx` on the forwards-pass, hence `NoRData`.
        dx = NoRData()

        return df, dx, dy
    end

    # Just return x (the CoDual) -- this propagates forwards the correct unique tangent
    # memory for `x`.
    return x, set_1_pullback!!
end

Let's unpack the above:

Memory Propagation

We stated at the top of this section that each unique address associated to differentiable data must have a unique tangent memory address associated to it. To see how this rule preserves this, consider the function

g(x::Vector{Float64}, y::Float64) = x, set_1!(x, y)

The output of g is a Tuple with the same Vector{Float64} in each element. Therefore, during AD, they must be associated to the same tangent address. Happily, simple by by returning x at the end of the rrule!! for set_1! we ensure that this happens.

The other field in a CoDual

In this example, the other field in the CoDual associated to x must contain a Vector{Float64}, which represents the tangent to x. We call this the fdata ("forwards data") associated to x. We didn't show it, but the fdata associated to y is NoFData ("no forwards data"), indicating that there is no additional data associated to y on the forwards-pass.

Why is this the case?

Summary

Note that this very simple function does not have a meaningful ChainRules.rrule counterpart because it mutates (modifies) x, and ChainRules.rrule does not support mutation.

Asides

Why Uniqueness of Type For Tangents / FData / RData?

Why does Tapir.jl insist that each primal type P be paired with a single tangent type T, as opposed to being more permissive. There are a few notable reasons:

To provide a precise interface. Rules pass fdata around on the forwards-pass and rdata on the reverse-pass – being able to make strong assumptions about the type of the fdata / rdata given the primal type makes implementing rules much easier in practice.

Conditional type stability. We wish to have a high degree of confidence that if the primal code is type-stable, then the AD code will also be. It is straightforward to construct type stable primal codes which have type-unstable forwards- and reverse-passes if you permit there to be more than one fdata / rdata type for a given primal. So while uniqueness is certainly not sufficient on its own to guarantee conditional type stability, it is probably necessary in general.

Test-case generation and coverage. There being a unique tangent / fdata / rdata type for each primal makes being confident that a given rule is being tested thoroughly much easier. For a given primal, rather than there being many possible input / output types to consider, there is just one.

This topic, in particular what goes wrong with permissive tangent type systems like those employed by ChainRules, deserves a more thorough treatment – hopefully someone will write something more expansive on this topic at some point.

Why Unique Memory Address