Safe Mode

The Problem

A major source of potential problems in AD systems is rules returning the wrong type of tangent / fdata / rdata for a given primal value. For example, if someone writes a rule like

function rrule!!(::CoDual{typeof(+)}, x::CoDual{<:Real}, y::CoDual{<:Real})
    plus_reverse_pass(dz::Real) = NoRData(), dz, dz
    return zero_fcodual(primal(x) + primal(y))
end

and calls

rrule(zero_fcodual(+), zero_fcodual(5.0), zero_fcodual(4f0))

then the type of dz on the reverse pass will be Float64 (assuming everything happens correctly), and this rule will return a Float64 as the rdata for y. However, the primal value of y is a Float32, so the appropriate tangent type is also Float32. This error may causes the reverse pass to fail loudly, but it may fail silently. It may cause an error much later in the reverse pass, making it hard to determine that the source of the error was the above rule. Worst of all, in some cases it could plausibly cause a segfault, which is more-or-less the worst kind of outcome possible.

The Solution

Check that the types of the fdata / rdata associated to arguments are exactly what tangent_type / fdata_type / rdata_type require upon entry to / exit from rules and pullbacks.

This is implemented via SafeRRule:

Tapir.SafeRRule — Type

SafeRRule(rule)

Construct a callable which is equivalent to rule, but inserts additional type checking. In particular:

check that the fdata in each argument is of the correct type for the primal

check that the fdata in the CoDual returned from the rule is of the correct type for the primal.

This happens recursively. For example, each element of a Vector{Any} is compared against each element of the associated fdata to ensure that its type is correct, as this cannot be guaranteed from the static type alone.

Some additional dynamic checks are also performed (e.g. that an fdata array of the same size as its primal).

Let rule return y, pb!!, then SafeRRule(rule) returns y, SafePullback(pb!!). SafePullback inserts the same kind of checks as SafeRRule, but on the reverse-pass. See the docstring for details.

Note: at any given point in time, the checks performed by this function constitute a necessary but insufficient set of conditions to ensure correctness. If you find that an error isn't being caught by these tests, but you believe it ought to be, please open an issue or (better still) a PR.

Note: this is a "safe mode" in the sense of operating systems. See e.g. this Wikipedia article: https://en.wikipedia.org/wiki/Safe_mode . Its purpose is to help with debugging, and should not be used when trying to differentiate code in general, as it decreases performance quite substantially in many cases.

You can straightforwardly enable it when building a rule via the safety_on kwarg in the following:

Tapir.build_rrule — Function

build_rrule(args...; safety_on=false)

Helper method. Only uses static information from args.

build_rrule(interp::TapirInterpreter{C}, sig_or_mi; safety_on=false) where {C}

Returns a DerivedRule which is an rrule!! for sig_or_mi in context C. See the docstring for rrule!! for more info.

If safety_on is true, then all calls to rules are replaced with calls to SafeRRules.

When using ADTypes.jl, you can choose whether or not to use it via the safe_mode kwarg:

julia> AutoTapir(safe_mode=false)
AutoTapir(safe_mode=false)

When Should You Use Safe Mode?

Only use safe_mode when debugging a problem. This is because is has substantial performance implications.