Tangents

As discussed in Representing Gradients, Mooncake requires that each "primal" type be associated to a unique "tangent" type, given by the function tangent_type. Moreover, we must be able to "split" a given tangent into its fdata ("forwards-data") and rdata ("reverse-data"), whose types are given by Mooncake.fdata_type and Mooncake.rdata_type respectively.

Very occassionally it may be necessary to specify your own tangent type. This is not an entirely trivial undertaking, as there is quite a lot of functionality that must be added to make it work properly. So, before diving in to add your own custom type, seriously consider whether it is worth the effort, and whether the default definition given by Mooncake are really inadequate for your use-case.

Testing Functionality

The interface is given in the form of three functions:

Mooncake.TestUtils.test_tangent_interface — Function

test_tangent_interface(rng::AbstractRNG, p; interface_only=false)

Verify that standard functionality for tangents runs, and is consistent. This function is the defacto formal definition of the "tangent interface" – if this function runs without error for a given value of p, then that p satisfies the tangent interface.

Extended Help

Verifies that the following functions are implemented correctly (as far as possible) for p / its type, and its tangents / their type:

tangent_type

zero_tangent

randn_tangent

TestUtils.has_equal_data

increment!!

set_to_zero!!

_add_to_primal

_diff

_dot

_scale

TestUtils.populate_address_map

In conjunction with the functions tested by test_tangent_splitting, these functions constitute a complete set of functions which must be applicable to p in order to ensure that it operates correctly in the context of reverse-mode AD. This list should be up to date at any given point in time, but the best way to verify that you've implemented everything is simply to run this function, and see whether it errors / produces a failing test.

Mooncake.TestUtils.test_tangent_splitting — Function

test_tangent_splitting(rng::AbstractRNG, p::P) where {P}

Verify that tangent splitting functionality associated to primal p works correctly. Ensure that test_tangent_interface runs for p before running these tests.

Extended Help

Verifies that the following functionality work correctly for p / its type / tangents:

Mooncake.fdata_type

Mooncake.rdata_type

Mooncake.fdata

Mooncake.rdata

Mooncake.uninit_fdata

Mooncake.tangent_type (binary method)

Mooncake.tangent (binary method)

Mooncake.TestUtils.test_rule_and_type_interactions — Function

test_rule_and_type_interactions(rng::AbstractRNG, p)

Check that a collection of standard functions for which we ought to have a working rrule for p work, and produce the correct answer. For example, the rrule!! for typeof should work correctly on any type, we should have a working rule for getfield for any struct-type, and we should have a rule for setfield! for any mutable struct type. See extended help for more info.

Extended Help

The purpose of this test is to ensure that, for any given p, the full range of primitive functions that ought to work on it, do indeed work on it.

This is one part of the interface where some care might be required. If, for some reason, it should never be the case that e.g. for a particular p, getfield should be called, then it may make no sense at all to run these tests. In such cases, the author of the type is responsible for knowing what they are doing. Please open an issue to discuss for your type if you are at all unsure what to do.

When defining a custom tangent type for P, the functions that you will need to pay attention to writing rules for are

Mooncake._new_

Mooncake.lgetfield

Mooncake.lsetfield!

In all cases, you may wish to consult the current implementations of rrule!! for these functions for inspiration regarding how you might implement them for your type.

You can call all three of these functions at once using

Mooncake.TestUtils.test_data — Function

test_data(rng::AbstractRNG, p::P)

Verify that all tangent / fdata / rdata functionality work properly for x. Furthermore, verify that all primitives listed in TestUtils.test_rule_and_type_interactions work correctly on x. This functionality is particularly useful if you are writing your own custom tangent / fdata / rdata types and want to be confident that you have implemented the functionality that you need in order to make these custom types work with all the rules written in Mooncake itself.

You should consult the docstrings for test_tangent_interface, test_tangent_splitting, and test_rule_and_type_interactions, in order to see what is required to satisfy the full tangent interface for p.

If all the tests in these functions pass, then you have satisfied the interface.

Interface

Below are the docstrings for each function tested by Mooncake.TestUtils.test_tangent_interface and Mooncake.TestUtils.test_tangent_splitting.

Mooncake.tangent_type — Function

tangent_type(P)

There must be a single type used to represents tangents of primals of type P, and it must be given by tangent_type(P).

Warning: this function assumes the effects :removable and :consistent. This is necessary to ensure good performance, but imposes precise constraints on your implementation. If adding new methods to tangent_type, you should consult the extended help of Base.@assume_effects to see what this imposes upon your implementation.

Extended help

The tangent types which Mooncake.jl uses are quite similar in spirit to ChainRules.jl. For example, tangent "vectors" for

Float64s are Float64s,

Vector{Float64}s are Vector{Float64}s, and

structs are other another (special) struct with field types specified recursively.

There are, however, some major differences. Firstly, while it is certainly true that the above tangent types are permissible in ChainRules.jl, they are not the uniquely permissible types. For example, ZeroTangent is also a permissible type of tangent for any of them, and Float32 is permissible for Float64. This is a general theme in ChainRules.jl – it intentionally declines to place restrictions on what type can be used to represent the tangent of a given type.

Mooncake.jl differs from this. It insists that each primal type is associated to a single tangent type. Furthermore, this type is always given by the function Mooncake.tangent_type(primal_type).

Consider some more worked examples.

Int

Int is not a differentiable type, so its tangent type is NoTangent:

julia> tangent_type(Int)
NoTangent

Tuples

The tangent type of a Tuple is defined recursively based on its field types. For example

julia> tangent_type(Tuple{Float64, Vector{Float64}, Int})
Tuple{Float64, Vector{Float64}, NoTangent}

There is one edge case to be aware of: if all of the field of a Tuple are non-differentiable, then the tangent type is NoTangent. For example,

julia> tangent_type(Tuple{Int, Int})
NoTangent

Structs

As with Tuples, the tangent type of a struct is, by default, given recursively. In particular, the tangent type of a struct type is Tangent. This type contains a NamedTuple containing the tangent to each field in the primal struct.

As with Tuples, if all field types are non-differentiable, the tangent type of the entire struct is NoTangent.

There are a couple of additional subtleties to consider over Tuples though. Firstly, not all fields of a struct have to be defined. Fortunately, Julia makes it easy to determine how many of the fields might possibly not be defined. The tangent associated to any field which might possibly not be defined is wrapped in a PossiblyUninitTangent.

Furthermore, structs can have fields whose static type is abstract. For example

julia> struct Foo
           x
       end

If you ask for the tangent type of Foo, you will see that it is

julia> tangent_type(Foo)
Tangent{@NamedTuple{x}}

Observe that the field type associated to x is Any. The way to understand this result is to observe that

x could have literally any type at runtime, so we know nothing about what its tangent type must be until runtime, and

we require that the tangent type of Foo be unique.

The consequence of these two considerations is that the tangent type of Foo must be able to contain any type of tangent in its x field. It follows that the fieldtype of the x field of Foos tangent must be Any.

Mutable Structs

The tangent type for mutable structs have the same set of considerations as structs. The only difference is that they must themselves be mutable. Consequently, we use a type called MutableTangent to represent their tangents. It is a mutable struct with the same structure as Tangent.

For example, if you ask for the tangent_type of

julia> mutable struct Bar
           x::Float64
       end

you will find that it is

julia> tangent_type(Bar)
MutableTangent{@NamedTuple{x::Float64}}

Primitive Types

We've already seen a couple of primitive types (Float64 and Int). The basic story here is that all primitive types require an explicit specification of what their tangent type must be.

One interesting case are Ptr types. The tangent type of a Ptr{P} is Ptr{T}, where T = tangent_type(P). For example

julia> tangent_type(Ptr{Float64})
Ptr{Float64}

tangent_type(F::Type, R::Type)::Type

Given the type of the fdata and rdata, F and R resp., for some primal type, compute its tangent type. This method must be equivalent to tangent_type(_typeof(primal)).

Mooncake.zero_tangent — Function

zero_tangent(x)

Returns the unique zero element of the tangent space of x. It is an error for the zero element of the tangent space of x to be represented by anything other than that which this function returns.

Internally, zero_tangent calls zero_tangent_internal, which handles different types of inputs differently. zero_tangent_internal has two variants:

For isbitstype types, zero_tangent_internal takes one argument.

Otherwise, zero_tangent_internal takes another argument which is an IdDict, which

handles both circular references and aliasing correctly.

zero_tangent(primal, fdata)

Equivalent to tangent(fdata, rdata(zero_tangent(primal))).

Mooncake.randn_tangent — Function

randn_tangent(rng::AbstractRNG, x::T) where {T}

Required for testing. Generate a randomly-chosen tangent to x. The design is closely modelled after zero_tangent.

Mooncake.TestUtils.has_equal_data — Function

has_equal_data(x, y; equal_undefs=true)

Determine if two objects x and y have equivalent data. If equal_undefs is true, undefined elements in arrays or unassigned fields in structs are considered equal.

The main logic is implemented in has_equal_data_internal, which is a recursive function that takes an additional visited dictionary to track visited objects and avoid infinite recursion in cases of circular references.

Mooncake.increment!! — Function

increment!!(x::T, y::T) where {T}

Add x to y. If ismutabletype(T), then increment!!(x, y) === x must hold. That is, increment!! will mutate x. This must apply recursively if T is a composite type whose fields are mutable.

Mooncake.set_to_zero!! — Function

set_to_zero!!(x)

Set x to its zero element (x should be a tangent, so the zero must exist).

Mooncake._add_to_primal — Function

_add_to_primal(p::P, t::T, unsafe::Bool=false) where {P, T}

Adds t to p, returning a P. It must be the case that tangent_type(P) == T.

If unsafe is true and P is a composite type, then _add_to_primal will construct a new instance of P by directly invoking the :new instruction for P, rather than attempting to use the default constructor for P. This is fine if you are confident that the new P constructed by adding t to p will always be a valid instance of P, but could cause problems if you are not confident of this.

This is, for example, fine for the following type:

struct Foo{T}
    x::Vector{T}
    y::Vector{T}
    function Foo(x::Vector{T}, y::Vector{T}) where {T}
        @assert length(x) == length(y)
        return new{T}(x, y)
    end
end

Here, the value returned by _add_to_primal will satisfy the invariant asserted in the inner constructor for Foo.

Mooncake._diff — Function

_diff(p::P, q::P) where {P}

Required for testing.

Computes the difference between p and q, which must be of the same type, P. Returns a tangent of type tangent_type(P).

Mooncake._dot — Function

_dot(t::T, s::T)::Float64 where {T}

Required for testing. Should be defined for all standard tangent types.

Inner product between tangents t and s. Must return a Float64. Always available because all tangent types correspond to finite-dimensional vector spaces.

Mooncake._scale — Function

_scale(a::Float64, t::T) where {T}

Required for testing. Should be defined for all standard tangent types.

Multiply tangent t by scalar a. Always possible because any given tangent type must correspond to a vector field. Not using * in order to avoid piracy.

Mooncake.TestUtils.populate_address_map — Function

populate_address_map(primal, tangent)

Constructs an empty AddressMap and calls populate_address_map_internal.

Mooncake.fdata_type — Function

fdata_type(T)

Returns the type of the forwards data associated to a tangent of type T.

Extended help

Rules in Mooncake.jl do not operate on tangents directly. Rather, functionality is defined to split each tangent into two components, that we call fdata (forwards-pass data) and rdata (reverse-pass data). In short, any component of a tangent which is identified by its address (e.g. a mutable structs or an Array) gets passed around on the forwards-pass of AD and is incremented in-place on the reverse-pass, while components of tangents identified by their value get propagated and accumulated only on the reverse-pass.

Given a tangent type T, you can find out what type its fdata and rdata must be with fdata_type(T) and rdata_type(T) respectively. A consequence of this is that there is exactly one valid fdata type and rdata type for each primal type.

Given a tangent t, you can get its fdata and rdata using f = fdata(t) and r = rdata(t) respectively. f and r can be re-combined to recover the original tangent using the binary version of tangent: tangent(f, r). It must always hold that

tangent(fdata(t), rdata(t)) === t

The need for all of this is explained in the docs, but for now it suffices to consider our running examples again, and to see what their fdata and rdata look like.

Int

Ints are non-differentiable types, so there is nothing to pass around on the forwards- or reverse-pass. Therefore

julia> fdata_type(tangent_type(Int)), rdata_type(tangent_type(Int))
(NoFData, NoRData)

Float64

The tangent type of Float64 is Float64. Float64s are identified by their value / have no fixed address, so

julia> (fdata_type(Float64), rdata_type(Float64))
(NoFData, Float64)

Vector{Float64}

The tangent type of Vector{Float64} is Vector{Float64}. A Vector{Float64} is identified by its address, so

julia> (fdata_type(Vector{Float64}), rdata_type(Vector{Float64}))
(Vector{Float64}, NoRData)

Tuple{Float64, Vector{Float64}, Int}

This is an example of a type which has both fdata and rdata. The tangent type for Tuple{Float64, Vector{Float64}, Int} is Tuple{Float64, Vector{Float64}, NoTangent}. Tuples have no fixed memory address, so we interogate each field on its own. We have already established the fdata and rdata types for each element, so we recurse to obtain:

julia> T = tangent_type(Tuple{Float64, Vector{Float64}, Int})
Tuple{Float64, Vector{Float64}, NoTangent}

julia> (fdata_type(T), rdata_type(T))
(Tuple{NoFData, Vector{Float64}, NoFData}, Tuple{Float64, NoRData, NoRData})

The zero tangent for (5.0, [5.0]) is t = (0.0, [0.0]). fdata(t) returns (NoFData(), [0.0]), where the second element is === to the second element of t. rdata(t) returns (0.0, NoRData()). In this example, t contains a mixture of data, some of which is identified by its value, and some of which is identified by its address, so there is some fdata and some rdata.

Structs

Structs are handled in more-or-less the same way as Tuples, albeit with the possibility of undefined fields needing to be explicitly handled. For example, a struct such as

julia> struct Foo
           x::Float64
           y
           z::Int
       end

has tangent type

julia> tangent_type(Foo)
Tangent{@NamedTuple{x::Float64, y, z::NoTangent}}

Its fdata and rdata are given by special FData and RData types:

julia> (fdata_type(tangent_type(Foo)), rdata_type(tangent_type(Foo)))
(Mooncake.FData{@NamedTuple{x::NoFData, y, z::NoFData}}, Mooncake.RData{@NamedTuple{x::Float64, y, z::NoRData}})

Practically speaking, FData and RData both have the same structure as Tangents and are just used in different contexts.

Mutable Structs

The fdata for a mutable structs is its tangent, and it has no rdata. This is because mutable structs have fixed memory addresses, and can therefore be incremented in-place. For example,

julia> mutable struct Bar
           x::Float64
           y
           z::Int
       end

has tangent type

julia> tangent_type(Bar)
MutableTangent{@NamedTuple{x::Float64, y, z::NoTangent}}

and fdata / rdata types

julia> (fdata_type(tangent_type(Bar)), rdata_type(tangent_type(Bar)))
(MutableTangent{@NamedTuple{x::Float64, y, z::NoTangent}}, NoRData)

Primitive Types

As with tangents, each primitive type must specify what its fdata and rdata is. See specific examples for details.

Mooncake.rdata_type — Function

rdata_type(T)

Returns the type of the reverse data of a tangent of type T.

Extended help

See extended help in fdata_type docstring.

Mooncake.fdata — Function

fdata(t)::fdata_type(typeof(t))

Extract the forwards data from tangent t.

Mooncake.rdata — Function

rdata(t)::rdata_type(typeof(t))

Extract the reverse data from tangent t.

Extended help

See extended help section of fdata_type.

Mooncake.uninit_fdata — Function

uninit_fdata(p)

Equivalent to fdata(uninit_tangent(p)).